+++ src/App.module.css
@@ -26,13 +26,13 @@ html {
 }
 
 .dark {
-transition: left 0.1s ease-in-out;
+  transition: left 0.1s ease-in-out;
   background: #282828;
   color: white;
 }
 
 .light {
-transition: left 0.1s ease-in-out;
+  transition: left 0.1s ease-in-out;
   background-color: white;
   color: black;
 }
