+++ src/Pages/MainPage.jsx
@@ -1,18 +1,15 @@
-import { useDispatch, useSelector } from "react-redux";
-import style from "./MainPage.module.css";
+import { useSelector } from "react-redux";
 import { Categories } from "../Components/Categories/Categories";
 import { DataUserForm } from "../Components/DataUserForm/DataUserForm";
 import { MainBanner } from "../Components/MainBanner/MainBanner";
 import { Sales } from "../Components/Sales/Sales";
-import { toggleTheme } from "../slices/themaSlice";
+import style from "./MainPage.module.css";
 
 export const MainPage = ({ toggleThemeHandler }) => {
-  const theme = useSelector(state => state.theme.theme)
-  const dispatch = useDispatch();
-
+  const theme = useSelector((state) => state.theme.theme);
 
   return (
-    <main className={`${theme === 'light' ? style.dark : style.light}`}>
+    <main className={`${theme === "light" ? style.dark : style.light}`}>
       <MainBanner toggleThemeHandler={toggleThemeHandler} />
       <Categories toggleThemeHandler={toggleThemeHandler} />
       <DataUserForm toggleThemeHandler={toggleThemeHandler} />
