+++ src/Pages/MainPage.module.css
@@ -7,4 +7,3 @@
   background: #282828;
   color: white;
 }
-
