+++ src/layout/Footer/Footer.jsx
@@ -1,14 +1,17 @@
 import React from "react";
-import style from "./Footer.module.css";
-import { GoogleMap } from "../../Components/GoogleMap/GoogleMap";
-import { ContactsCards } from "../../Components/ContactsCards/ContactsCards";
 import { useSelector } from "react-redux";
+import { ContactsCards } from "../../Components/ContactsCards/ContactsCards";
+import { GoogleMap } from "../../Components/GoogleMap/GoogleMap";
+import style from "./Footer.module.css";
 
 export const Footer = () => {
-  const theme = useSelector((state) => state.theme.theme)
+  const theme = useSelector((state) => state.theme.theme);
   /**Массив объектов создан для облегчения добавления и изменения данных в случае если это потребуется */
   const contactsCardData = [
-    { title: "Phone", content: <a href="tel:+499999999999">+49 999 999 99 99</a> },
+    {
+      title: "Phone",
+      content: <a href="tel:+499999999999">+49 999 999 99 99</a>,
+    },
     { title: "Social", content: "" },
     {
       title: "Address",
@@ -17,16 +20,23 @@ export const Footer = () => {
     { title: "Working Hours", content: "24 hours a day" },
   ];
   return (
-    <footer className={`${style.footerWrapper} ${theme === 'light' ? style.dark : style.light}`}>
-      <h1 className={`${style.footerTitle} ${theme === 'light' ? style.dark : style.light}`}>Contact</h1>
+    <footer
+      className={`${style.footerWrapper} ${
+        theme === "light" ? style.dark : style.light
+      }`}
+    >
+      <h1
+        className={`${style.footerTitle} ${
+          theme === "light" ? style.dark : style.light
+        }`}
+      >
+        Contact
+      </h1>
       <div className={style.footerCards}>
         {contactsCardData.map(
           /** с помощью метода map перебираем карточки и рендерим ContactCards для каждого элемента массива */
           (card, index) => (
-            <ContactsCards
-              cardData={card}
-              key={index}
-            /> /**также можно здесь использовать item.id в качестве ключа для каждого элемента списка */
+            <ContactsCards cardData={card} key={index} />
           )
         )}
       </div>
@@ -34,5 +44,3 @@ export const Footer = () => {
     </footer>
   );
 };
-
-/**Код представляет собой React компонент под названием Footer, который рендерит футер веб-страницы. В футере отображаются контактные данные (телефон, социальные сети, адрес и часы работы) с использованием компонента ContactsCards и карта с использованием компонента GoogleMap */
