+++ src/slices/breadcrumbsSlice.js
@@ -1,18 +1,18 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
-    breadcrumbsList: [],
-  };
-  
-export const breadcrumbsSlice = createSlice({
-    name: 'breadcrumbs',
-  initialState,
-    reducers: {
-        setBreadcrumbs: (state, action) => {
-            state.breadcrumbsList = action.payload
-        }
-    }
-})
+  breadcrumbsList: [],
+};
 
-export const {setBreadcrumbs} = breadcrumbsSlice.actions
-export default breadcrumbsSlice.reducer
+export const breadcrumbsSlice = createSlice({
+  name: "breadcrumbs",
+  initialState,
+  reducers: {
+    setBreadcrumbs: (state, action) => {
+      state.breadcrumbsList = action.payload;
+    },
+  },
+});
+
+export const { setBreadcrumbs } = breadcrumbsSlice.actions;
+export default breadcrumbsSlice.reducer;
