+++ src/slices/themaSlice.js
@@ -1,19 +1,19 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
-    theme: localStorage.getItem('theme') || 'dark',
-}
+  theme: localStorage.getItem("theme") || "dark",
+};
 
 const themeSlice = createSlice({
-    name: 'theme',
-    initialState,
-    reducers: {
-        toggleTheme: (state) => {
-            state.theme = state.theme === 'light' ? 'dark' : 'light',
-                localStorage.setItem('theme', state.theme);
-        }
-    }
-})
+  name: "theme",
+  initialState,
+  reducers: {
+    toggleTheme: (state) => {
+      (state.theme = state.theme === "light" ? "dark" : "light"),
+        localStorage.setItem("theme", state.theme);
+    },
+  },
+});
 
 export const { toggleTheme } = themeSlice.actions;
 export default themeSlice.reducer;
